Restructure unit tests folder
